# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KPPDF CRM API
  description: REST API для системы управления коммерческими предложениями и производством
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development
security:
  - BearerAuth: []
paths:
...
  /upload:
    post:
      summary: Загрузить файл
      tags: [System]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses: { '200': { description: Файл загружен } }
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Cookie
      description: JWT token в cookie (accessToken)
...